requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dominate==2.8.0
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mysqlclient==2.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.5
visitor==0.1.3
Werkzeug==2.3.7
